Fixed shortcuts.
- Node can be configured/reconfigured via shortcuts.
- Node can be launched via shortcuts w/o crashing.
- Todo start/stop node as a service.
