+++ install/windows/TestNodeLaunch/Program.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace TestNodeLaunch
+{
+    class Program
+    {
+        static void Main(string[] args)
+        {
+            Console.WriteLine("Hello World!");
+        }
+    }
+}
