+++ install/windows/cypnode_installer.sln
@@ -6,9 +6,12 @@ MinimumVisualStudioVersion = 10.0.40219.1
 Project("{930C7802-8A8C-48F9-8165-68863BCCD9DD}") = "cypnode_installer", "cypnode_installer\cypnode_installer.wixproj", "{162C4BB5-4DB8-4503-8AA8-66BCFDE83EE1}"
 	ProjectSection(ProjectDependencies) = postProject
 		{9ED3C22D-B902-43BC-AF4E-BDF2EC3305C9} = {9ED3C22D-B902-43BC-AF4E-BDF2EC3305C9}
+		{4D4ACAE8-35BB-42D9-A1B9-42324601F127} = {4D4ACAE8-35BB-42D9-A1B9-42324601F127}
 	EndProjectSection
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "service", "service\service.csproj", "{9ED3C22D-B902-43BC-AF4E-BDF2EC3305C9}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestNodeLaunch", "TestNodeLaunch\TestNodeLaunch.csproj", "{4D4ACAE8-35BB-42D9-A1B9-42324601F127}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -42,6 +45,18 @@ Global
 		{9ED3C22D-B902-43BC-AF4E-BDF2EC3305C9}.Release|x64.Build.0 = Release|Any CPU
 		{9ED3C22D-B902-43BC-AF4E-BDF2EC3305C9}.Release|x86.ActiveCfg = Release|Any CPU
 		{9ED3C22D-B902-43BC-AF4E-BDF2EC3305C9}.Release|x86.Build.0 = Release|Any CPU
+		{4D4ACAE8-35BB-42D9-A1B9-42324601F127}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4D4ACAE8-35BB-42D9-A1B9-42324601F127}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4D4ACAE8-35BB-42D9-A1B9-42324601F127}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{4D4ACAE8-35BB-42D9-A1B9-42324601F127}.Debug|x64.Build.0 = Debug|Any CPU
+		{4D4ACAE8-35BB-42D9-A1B9-42324601F127}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{4D4ACAE8-35BB-42D9-A1B9-42324601F127}.Debug|x86.Build.0 = Debug|Any CPU
+		{4D4ACAE8-35BB-42D9-A1B9-42324601F127}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4D4ACAE8-35BB-42D9-A1B9-42324601F127}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4D4ACAE8-35BB-42D9-A1B9-42324601F127}.Release|x64.ActiveCfg = Release|Any CPU
+		{4D4ACAE8-35BB-42D9-A1B9-42324601F127}.Release|x64.Build.0 = Release|Any CPU
+		{4D4ACAE8-35BB-42D9-A1B9-42324601F127}.Release|x86.ActiveCfg = Release|Any CPU
+		{4D4ACAE8-35BB-42D9-A1B9-42324601F127}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
